web-semantique-main/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
rdflib==7.0.0
SPARQLWrapper==2.0.0
requests==2.31.0
google-generativeai==0.3.1
python-dotenv==1.0.0
